フォルダ構成変更とコンポーネントの名称変更
