--- src/App.js
@@ -1,20 +1,0 @@
-import React from 'react'
-
-import Footer from './Footer'
-import Header from './Header'
-import Router from './Router'
-
-
-class App extends React.Component {
-    render() {
-        return (
-            <>
-                <Header />
-                <Router />
-                <Footer />
-            </>
-        )
-    }
-}
-
-export default App
